fix tainted input in backup_cfs_database
We can call this module also now over the API through
pvedaemon/pveproxy which have tainting checks on.

Thus we need to untaint the "read existing backups" inputs,
as else this errors out here.

Only triggers when over 10 backups existed already, so this does
not triggers really often in the real world.

+++ data/PVE/Cluster.pm
@@ -1762,6 +1762,7 @@ my $backup_cfs_database = sub {
 
     if ((my $count = scalar(@$backups)) > $maxfiles) {
 	foreach my $f (@$backups[$maxfiles..$count-1]) {
+	    next if $f !~ m/^(\S+)$/; # untaint
 	    print "delete old backup '$1'\n";
 	    unlink $1;
 	}
